Add a note about the origins of Linode, which traces back to 1999
TheShore.net was a ColdFusion hosting company started in 1999. Its core
components and codebase became the core of Linode, in 2003.

Fun-fact: parts of the core billing system still in-use in 2015 were originally
written for TheShore.net in 1999-2000. They may still be in-use today, but I
can't speak to that.

Full Disclosure: I was a developer / administrator there in a past life.

Signed-off-by: Tim Heckman <t@heckman.io>